+++ react-form/src/App.css
@@ -2,53 +2,3 @@
   max-width: 1280px;
   margin: 0 auto;
 }
-
-.Form{
-  padding: 1%;
-  border: 1px solid #000;
-  border-radius: 10px;
-  width: 40%;
-  /* background-color: #a5abb1; */
-  background: linear-gradient(90deg,rgb(148, 149, 160) 6%,rgb(172, 173, 185) 34%,rgb(193, 194, 207) 100%);
-  box-shadow:
-  0 1px 4px rgba(0, 0, 0, .3),
-  -23px 0 20px -23px rgba(0, 0, 0, .8),
-  23px 0 20px -23px rgba(0, 0, 0, .8),
-  0 0 40px rgba(0, 0, 0, .1) inset;
-  margin: 0 auto;
-}
-
-.Form div input{
-  cursor: pointer;
-  background-color: rgb(241, 235, 235);
-  width: 99%;
-  border: 1px solid #000;
-  color: rgb(0, 0, 0);
-  border-radius: 10px;
-  outline: none;
-  font-size: 18px;
-  text-align: center;
-  margin: 1% auto;
-}
-
-.Selrct-form{
-  cursor: pointer;
-  background-color: rgb(241, 235, 235);
-  width: 99%;
-  border: 1px solid #000;
-  color: rgb(0, 0, 0);
-  border-radius: 10px;
-  font-size: 18px;
-}
-
-.label-pol{
-  display: flex;
-}
-
-.error-input{
-  border: 1px solid red !important;
-}
-
-.error-black{
-  border: 1px solid black !important;
-}
