+++ react-form/src/components/Formcarts.css
@@ -1,0 +1,24 @@
+.img-cart{
+  width: 100px;
+  height: 100px;
+  border-radius: 50%;
+  background: rgb(175, 62, 62);
+}
+
+.cart-wrapper{
+display: inline-block;
+background: rgb(129, 126, 126);
+padding: 10px;
+border-radius: 10px;
+margin: 10px;
+text-align: center;
+}
+
+.cart-wrapper p{
+padding: 0;
+margin: 0;
+}
+
+.cart-wrapper p span{
+  font-weight: bold;
+  }
