+++ react-form/src/components/Header/Header.css
@@ -1,0 +1,26 @@
+.Header{
+  display: flex;
+  width: 100%;
+  height: 80px;
+  background: rgb(82, 78, 78);
+  align-items: center;
+}
+.Header-ul{
+  display: flex;
+  list-style: none;
+}
+
+.Header-ul li{
+  margin-right: 15px;
+}
+
+.active-link {
+  color: var(---color-active);
+  cursor: default;
+}
+
+li > a{
+  color: rgb(255, 255, 255);
+}
+
+
