+++ react-form/src/components/Header/Header.tsx
@@ -1,0 +1,34 @@
+import React from 'react';
+import { NavLink } from 'react-router-dom';
+import '../Header/Header.css';
+
+const Header = () => {
+  const setActive = ({ isActive }: { isActive: boolean }): string =>
+    isActive ? 'active-link' : '';
+
+  return (
+    <div className="Header">
+      <nav>
+        <ul className="Header-ul">
+          <li>
+            <NavLink to="/" className={setActive} style={{ textDecoration: 'none' }}>
+              Home
+            </NavLink>
+          </li>
+          <li>
+            <NavLink to="/about" className={setActive} style={{ textDecoration: 'none' }}>
+              About Us
+            </NavLink>
+          </li>
+          <li>
+            <NavLink to="/form" className={setActive} style={{ textDecoration: 'none' }}>
+              Form
+            </NavLink>
+          </li>
+        </ul>
+      </nav>
+    </div>
+  );
+};
+
+export default Header;
