+++ react-form/src/components/Сarts/Cart.css
@@ -1,0 +1,16 @@
+.Carts{
+  display: flex;
+  justify-content: space-around;
+  flex-wrap: wrap;
+}
+
+.Cart{
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.Cart img{
+  height: 200px;
+  cursor: pointer;
+}
