+++ react-form/src/page/AboutUs.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class AboutUs extends React.Component {
+  render() {
+    return (
+      <div>
+        <h2>about us</h2>
+      </div>
+    );
+  }
+}
+
+export default AboutUs;
