+++ react-form/src/page/Error.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class Error extends React.Component {
+  render() {
+    return (
+      <div>
+        <h2>error</h2>
+      </div>
+    );
+  }
+}
+
+export default Error;
